pyproject.toml
[tool.poetry]
name = "dp_entrypoint"
version = "0.1.0"
description = ""
authors = ["DiegoPomares"]

[tool.poetry.dependencies]
python = "^3.6"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
